# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management API",
        "description": "API REST pour la gestion d'une bibliothèque - Système de gestion des auteurs et livres",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Liste des auteurs",
                "description": "Récupère la liste paginée des auteurs avec recherche et tri",
                "operationId": "b1de4facd6962392c6fde61f8a7b097a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans prénom/nom",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Victor"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "first_name",
                                "last_name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction du tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des auteurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Author"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Créer un nouvel auteur",
                "description": "Crée un nouvel auteur avec prénom et nom",
                "operationId": "ff9d12ef46fc5034a0d4adf9d10b6de3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Victor"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Hugo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Auteur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Afficher un auteur",
                "description": "Récupère les détails d'un auteur spécifique",
                "operationId": "35c1b6d7002827ab95f68425b845db97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'auteur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'auteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Auteur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Author]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Mettre à jour un auteur",
                "description": "Met à jour les informations d'un auteur existant",
                "operationId": "8797e8c11aa3838017483c14703f088c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'auteur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Victor"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Hugo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auteur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Auteur non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Supprimer un auteur",
                "description": "Supprime un auteur (impossible s'il a des livres)",
                "operationId": "c56695bdfa8aa06b4a00f609c64bdc40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'auteur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auteur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Auteur supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Suppression impossible (auteur a des livres)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Impossible de supprimer cet auteur car il a des livres associés."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "AUTHOR_HAS_BOOKS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Auteur non trouvé"
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Liste des livres",
                "description": "Récupère la liste paginée des livres avec recherche, filtres et tri",
                "operationId": "9017a3fa89d6636329067c858b27a6fa",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans titre ou nom d'auteur",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Misérables"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Filtrer par ID d'auteur",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Prix minimum",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 10
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Prix maximum",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 50
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "price",
                                "publication_date",
                                "author_name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction du tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des livres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Créer un nouveau livre",
                "description": "Crée un nouveau livre avec titre, prix, date et auteur",
                "operationId": "72f82e1cd6f08112323f87053277ca99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "publication_date",
                                    "author_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Les Misérables"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "maximum": 999.99,
                                        "minimum": 0,
                                        "example": 25.99
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1862-01-01"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Afficher un livre",
                "description": "Récupère les détails d'un livre spécifique",
                "operationId": "123856ec83e3466396b979a43eb1a15f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du livre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Mettre à jour un livre",
                "description": "Met à jour les informations d'un livre existant",
                "operationId": "2fc15aa806ac989f4d2b201b5952714d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Les Misérables"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "maximum": 999.99,
                                        "minimum": 0,
                                        "example": 25.99
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1862-01-01"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Supprimer un livre",
                "description": "Supprime un livre de la bibliothèque",
                "operationId": "eda12c526d0ccacaecea5f3d8f15d6df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Livre supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Author",
                "description": "Modèle représentant un auteur",
                "properties": {
                    "id": {
                        "description": "Identifiant unique de l'auteur",
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Prénom de l'auteur",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Victor"
                    },
                    "last_name": {
                        "description": "Nom de famille de l'auteur",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Hugo"
                    },
                    "full_name": {
                        "description": "Nom complet de l'auteur",
                        "type": "string",
                        "example": "Victor Hugo"
                    },
                    "books_count": {
                        "description": "Nombre de livres de l'auteur",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Book": {
                "title": "Book",
                "description": "Modèle représentant un livre",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du livre",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Titre du livre",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Les Misérables"
                    },
                    "price": {
                        "description": "Prix du livre en euros",
                        "type": "number",
                        "format": "float",
                        "maximum": 999.99,
                        "minimum": 0,
                        "example": 19.99
                    },
                    "publication_date": {
                        "description": "Date de publication",
                        "type": "string",
                        "format": "date",
                        "example": "1862-04-03"
                    },
                    "author_id": {
                        "description": "ID de l'auteur",
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer token for authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}